requirements.txt
Flask>=3.0,<4
yt-dlp>=2023.11
openai-whisper>=20230314
torch>=2.0.0